Fixed object save method params documentation
The documentation for the params required for Object's save method describes 1 param (options), but you cannot actually pass just options as an object to the save method. The correct use of the save method is described with examples in the documentation, but the additional 2 params are not documented. As the save method does accept 3 arguments (in the code as value: function (arg1, arg2, arg3) {), these params need to be added to the documentation and not just mentioned in the examples.