Support global request batch size (#1053)
Closes: https://github.com/parse-community/Parse-SDK-JS/issues/1045

Currently the /batch endpoint send multiple requests in the with the number of objects equal to batchSize (default 20). I believe batchSize was added to reduce server load during the parse.com days. Having batchSize will make result in a certain amount of overhead for your client.

This PR will allow you to decrease the number of request which may result in saving resources.