Make ParseQuery.map respect returned promises
This function is documented as allowing the callback to return a
promise, and when the callback returns a promise we should wait for that
promise to resolve before continuing, and stop iteration if the promise
rejects.

If the callback returned a promise, the previous implementation would
return an array of promises, and run the callback on all of them. The
new implementation wraps the returned value in `Promise.resolve`,
coercing it into a promise, so that we can safely extract a value from
it before pushing the value into the results array. Returning this
promise in the `.each()` callback also provides the behavior of waiting
/ stopping based on the resolution of a promise from the callback.