Make ParseQuery.reduce respect returned promises
This function is documented as allowing the callback to return a
promise, and when the callback returns a promise we should wait for that
promise to resolve before continuing, and stop iteration if the promise
rejects.

The previous implementation was a bit naive in that it would load the
entire set of objects matching the query into memory, and then call
`.reduce()` on it. The new implementation avoids this by applying the
callback on each object in turn. This lets us then return the result of
the callback to the `.each()` call, which handles waiting for promise
resolution / rejection.

Things here get a little tricky in that the behavior of `reduce` is
special when no initial value is provided: we take the first item, don't
pass it into the callback, and treat that item as the initial value. We
need to keep that treatment, which entails a check for a special case:
we're looking at the first item AND the initial value was undefined.