Make ParseQuery.filter respect returned promises
This function is documented as allowing the callback to return a
promise, and when the callback returns a promise we should wait for that
promise to resolve before continuing, and stop iteration if the promise
rejects.

When the callback returns a promise, the old behavior would have been to
push the object into the results array, because a promise is a truthy
value. Now we wait for the promise to resolve before checking the value.