Update CloudCode.js
Hi all,

Not sure if my formatting is perfect but closes #1176 (removes pre. SDK 2.0 success functions), and adds documentation for LiveQuery triggers.

Also, for me, when I google "Parse Javascript SDK", it takes me to v1.11.0. Is there any way we can add "This SDK is outdated" or some other warning to point users to the latest SDK?

Thank you!