Add further complexity to test cases when documents changed
A bug has been identified when we attempt to increment a value in an embedded
document.  This could potentially be missed by the current tests because they
only have a document with a single value and that is the value that is updated.

This commit adds a second property to the embedded document and ensures it is
still present after the various function calls