Bug/server 6899/embedded document increment (#1219)
* Add further complexity to test cases when documents changed

A bug has been identified when we attempt to increment a value in an embedded
document.  This could potentially be missed by the current tests because they
only have a document with a single value and that is the value that is updated.

This commit adds a second property to the embedded document and ensures it is
still present after the various function calls

* Add tests that demonstrate the issue with incrementing embedded doc

Once the object is saved, the response from the save only returns the updated
properties on the embedded document.  This is then used to REPLACE the embedded
document instead of being merged with it.  This is incorrect if the operation
was an Increment/Decrement/Add

* Attempt to resolve embedded document increment bug

This change resolves the system when we use it in our code... however it does
NOT result in the new tests that were created passing.  I am a little confused
by that.

* Adjust handleSaveResponse conditional handling.

Only certain types of objects should be merged using Javascript spread operator.
Previous attempt was including Dates... which fails

* _handleSaveResponse: Merge response value with original attributes for nested objects

Co-authored-by: Diamond Lewis <findlewis@gmail.com>
Co-authored-by: Robby Helms <robby@pocketprep.com>